pyproject.toml
[tool.poetry]
name = "cartola-2023"
version = "0.1.0"
description = ""
authors = ["Gabriel Mizuno <gabrielmizuno@gmail.com>"]
readme = "README.md"
packages = [{include = "cartola_2023"}]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
pendulum = "^2.1.2"
pyarrow = "^11.0.0"
dataclasses-json = "^0.5.7"
google-cloud-bigquery = "^3.9.0"
cartola-project = {git = "https://github.com/GMizuno/cartola_project"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
